--- src/main/java/com/fours/tolevelup/controller/api/CommentController.java
@@ -1,12 +1,0 @@
-package com.fours.tolevelup.controller.api;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/comments/")
-public class CommentController {
-}
