+++ src/main/java/com/fours/tolevelup/service/FeedService.java
@@ -3,9 +3,9 @@ package com.fours.tolevelup.service;
 
 import com.fours.tolevelup.exception.ErrorCode;
 import com.fours.tolevelup.exception.TluApplicationException;
-import com.fours.tolevelup.model.FeedDTO;
-import com.fours.tolevelup.model.UserDTO;
-import com.fours.tolevelup.model.UserDTO.feedUserData;
+import com.fours.tolevelup.service.dto.FeedDTO;
+import com.fours.tolevelup.service.dto.UserDTO;
+import com.fours.tolevelup.service.dto.UserDTO.feedUserData;
 import com.fours.tolevelup.model.entity.User;
 import com.fours.tolevelup.model.entity.UserCharacter;
 import com.fours.tolevelup.repository.FollowRepository;
