+++ src/main/java/com/fours/tolevelup/service/MissionService.java
@@ -3,26 +3,18 @@ package com.fours.tolevelup.service;
 
 import com.fours.tolevelup.exception.ErrorCode;
 import com.fours.tolevelup.exception.TluApplicationException;
-import com.fours.tolevelup.model.MissionDTO;
-import com.fours.tolevelup.model.MissionStatus;
-import com.fours.tolevelup.model.ThemeType;
+import com.fours.tolevelup.service.dto.MissionDTO;
 import com.fours.tolevelup.model.entity.Character;
 import com.fours.tolevelup.model.entity.Mission;
 import com.fours.tolevelup.model.entity.MissionLog;
-import com.fours.tolevelup.model.entity.Theme;
 import com.fours.tolevelup.model.entity.User;
 import com.fours.tolevelup.model.entity.UserCharacter;
 import com.fours.tolevelup.repository.character.CharacterRepository;
 import com.fours.tolevelup.repository.character.UserCharacterRepository;
 import com.fours.tolevelup.repository.MissionRepository;
 import com.fours.tolevelup.repository.MissionLogRepository;
-import com.fours.tolevelup.repository.ThemeRepository;
 import com.fours.tolevelup.repository.ThemeExpRepository;
 import com.fours.tolevelup.repository.UserRepository;
-import java.sql.Date;
-import java.sql.Timestamp;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
