+++ src/main/java/com/fours/tolevelup/service/UserService.java
@@ -4,9 +4,9 @@ import com.fours.tolevelup.controller.request.UserUpdateRequest;
 import com.fours.tolevelup.controller.response.UserResponse;
 import com.fours.tolevelup.exception.ErrorCode;
 import com.fours.tolevelup.exception.TluApplicationException;
-import com.fours.tolevelup.model.AlarmDTO;
-import com.fours.tolevelup.model.ThemeExpDTO;
-import com.fours.tolevelup.model.UserDTO;
+import com.fours.tolevelup.service.dto.AlarmDTO;
+import com.fours.tolevelup.service.dto.ThemeExpDTO;
+import com.fours.tolevelup.service.dto.UserDTO;
 import com.fours.tolevelup.model.entity.Alarm;
 import com.fours.tolevelup.model.entity.Character;
 import com.fours.tolevelup.model.entity.Theme;
