+++ src/main/java/com/fours/tolevelup/service/dto/FeedDTO.java
@@ -1,16 +1,11 @@
-package com.fours.tolevelup.model;
+package com.fours.tolevelup.service.dto;
 
 
-import com.fours.tolevelup.model.FeedDTO.CharacterData;
-import com.fours.tolevelup.model.entity.Comment;
 import com.fours.tolevelup.service.character.CharacterDTO;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import org.springframework.data.domain.Page;
 
-import java.sql.Date;
-import java.sql.Timestamp;
 import java.util.List;
 
 @Getter
