+++ src/main/java/com/fours/tolevelup/service/dto/MissionDTO.java
@@ -1,14 +1,10 @@
-package com.fours.tolevelup.model;
+package com.fours.tolevelup.service.dto;
 
-import com.fours.tolevelup.model.entity.Mission;
+import com.fours.tolevelup.model.MissionStatus;
 import com.fours.tolevelup.model.entity.MissionLog;
-import com.fours.tolevelup.model.entity.Theme;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
-
-import java.util.List;
 
 
 @Getter
