+++ src/main/java/com/fours/tolevelup/service/dto/StatsDTO.java
@@ -1,7 +1,6 @@
-package com.fours.tolevelup.model;
+package com.fours.tolevelup.service.dto;
 
 
-import com.fours.tolevelup.model.entity.Theme;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
