+++ src/main/java/com/fours/tolevelup/service/dto/ThemeExpDTO.java
@@ -1,12 +1,10 @@
-package com.fours.tolevelup.model;
+package com.fours.tolevelup.service.dto;
 
 import com.fours.tolevelup.model.entity.Theme;
 import com.fours.tolevelup.model.entity.ThemeExp;
 import com.fours.tolevelup.model.entity.User;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
 
 @Getter
